tests: split tests in files
