+++ src/config/test.ts
@@ -1,0 +1,7 @@
+interface ITestConfig {
+  baseURL: string;
+}
+
+export const testConfig: ITestConfig = {
+  baseURL: `http://localhost:${process.env.SERVER_PORT}/graphql`,
+};
