+++ src/test/index.spec.ts
@@ -1,37 +1,14 @@
 import '@config/env';
+import { testConfig } from '@config/test';
 import { expect } from 'chai';
-import { beforeEach } from 'mocha';
-import { JWTProvider } from 'src/typeorm/entities/users/providers/jwt-provider/jwt-provider';
 import { describe, it } from 'mocha';
 import { runServer } from 'src/server';
-import { HashProvider } from 'src/typeorm/entities/users/providers/hash-provider/hash-provider';
-import { User } from 'src/typeorm/entities/users/user-entity';
 import supertest from 'supertest';
-import { getRepository, Repository } from 'typeorm';
-import { QueryLoginMutation } from './request-builder';
 
-let baseURL: string;
-let usersRepository: Repository<User>;
-let hashProvider: HashProvider;
-let jwtProvider: JWTProvider;
-let user: User;
-
-const expectedUserData = {
-  name: 'username1',
-  email: 'useremail1@provider.com',
-  birthDate: 'userbirthDate1',
-  cpf: 'usercpf1',
-};
+const baseURL = testConfig.baseURL;
 
 before(async () => {
   await runServer();
-
-  usersRepository = getRepository(User);
-
-  hashProvider = new HashProvider();
-  jwtProvider = new JWTProvider();
-
-  baseURL = `http://localhost:${process.env.SERVER_PORT}/graphql`;
 });
 
 describe('Testing GraphQL - Hello', () => {
@@ -48,205 +25,5 @@ describe('Testing GraphQL - Hello', () => {
   });
 });
 
-describe('E2E GraphQL - Mutation - User', () => {
-  beforeEach(async () => {
-    const password = await hashProvider.generate(`userpassword1`);
-    user = await usersRepository.save({
-      name: `username1`,
-      email: `useremail1@provider.com`,
-      password,
-      birthDate: `userbirthDate1`,
-      cpf: `usercpf1`,
-    });
-  });
-
-  afterEach(async () => await usersRepository.delete({}));
-
-  it('Should perform a successfull login returning a valid jwt token', async () => {
-    const request = QueryLoginMutation({
-      email: 'useremail1@provider.com',
-      password: 'userpassword1',
-      rememberMe: false,
-    });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-    const response = body.data.login;
-
-    const verifiedToken = await jwtProvider.verify(response.token);
-
-    expect(response)
-      .to.have.property('user')
-      .to.be.deep.eq({
-        ...expectedUserData,
-        id: user.id.toString(),
-      });
-
-    expect(response).to.have.property('token').which.is.a('string');
-    expect(verifiedToken.payload.userId).to.be.eq(user.id);
-  });
-
-  it('Should validate jwt token after expiration time (not remember me)', async () => {
-    const request = QueryLoginMutation({
-      email: 'useremail1@provider.com',
-      password: 'userpassword1',
-      rememberMe: false,
-    });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-    const response = body.data.login;
-
-    const verifiedToken = await jwtProvider.verify(response.token);
-
-    expect(response)
-      .to.have.property('user')
-      .to.be.deep.eq({
-        ...expectedUserData,
-        id: user.id.toString(),
-      });
-
-    const currentTimestamp = new Date();
-    currentTimestamp.setMilliseconds(0);
-
-    const expectExpiration = currentTimestamp;
-
-    const minutesToExpire = (verifiedToken.exp - verifiedToken.iat) / 60;
-
-    expectExpiration.setMinutes(currentTimestamp.getMinutes() + minutesToExpire);
-
-    const tokenExpires = new Date(verifiedToken.exp * 1000);
-
-    const expiresIn = +expectExpiration === +tokenExpires;
-    const isTokenExpired = currentTimestamp > tokenExpires;
-
-    expect(isTokenExpired).to.be.eq(false);
-    expect(expiresIn).to.be.eq(true);
-  });
-
-  it('Should validate jwt token after expiration time (remember me)', async () => {
-    const request = QueryLoginMutation({
-      email: 'useremail1@provider.com',
-      password: 'userpassword1',
-      rememberMe: true,
-    });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-    const response = body.data.login;
-
-    const verifiedToken = await jwtProvider.verify(response.token);
-
-    expect(response)
-      .to.have.property('user')
-      .to.be.deep.eq({
-        ...expectedUserData,
-        id: user.id.toString(),
-      });
-
-    const currentTimestamp = new Date();
-    currentTimestamp.setMilliseconds(0);
-
-    const expectExpiration = currentTimestamp;
-
-    const minutesToExpire = (verifiedToken.exp - verifiedToken.iat) / 60;
-
-    expectExpiration.setMinutes(currentTimestamp.getMinutes() + minutesToExpire);
-
-    const tokenExpires = new Date(verifiedToken.exp * 1000);
-
-    const expiresIn = +expectExpiration === +tokenExpires;
-    const isTokenExpired = currentTimestamp > tokenExpires;
-
-    expect(isTokenExpired).to.be.eq(false);
-    expect(expiresIn).to.be.eq(true);
-  });
-
-  it('Should not validate jwt token after expiration time', async () => {
-    const request = QueryLoginMutation({
-      email: 'useremail1@provider.com',
-      password: 'userpassword1',
-      rememberMe: false,
-    });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-    const response = body.data.login;
-
-    const verifiedToken = await jwtProvider.verify(response.token);
-
-    const tokenExpires = new Date(verifiedToken.exp * 1000);
-    const afterExpires = new Date();
-    afterExpires.setMilliseconds(0);
-
-    const expectExpiration = afterExpires;
-    expectExpiration.setMinutes(afterExpires.getMinutes() - 1);
-    const isTokenExpired = tokenExpires > afterExpires;
-
-    expect(isTokenExpired).to.be.eql(true);
-  });
-
-  it('Should not login without email or password.', async () => {
-    const request = QueryLoginMutation({ email: '', password: '' });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-
-    expect(body.errors).to.have.lengthOf.least(1);
-
-    expect(body.errors[0]).to.have.property('message').eql('Unauthorized. Possible invalid credentials.');
-
-    expect(body.errors[0].extensions.exception)
-      .to.have.property('additionalInfo')
-      .to.be.eql('Email or Password is null.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('code').eq(401);
-
-    expect(body).to.have.nested.property('data').to.be.null;
-  });
-
-  it('Should not login with invalid email.', async () => {
-    const request = QueryLoginMutation({ email: 'invalid Email Input', password: 'invalid Password' });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-
-    expect(body.errors).to.have.lengthOf.least(1);
-
-    expect(body.errors[0]).to.have.property('message').eql('Unauthorized. Possible invalid credentials.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('additionalInfo').to.be.eql('Invalid email format.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('code').eq(401);
-
-    expect(body).to.have.nested.property('data').to.be.null;
-  });
-
-  it('Should not login if user does not exist.', async () => {
-    const request = QueryLoginMutation({ email: 'InvalidUser@Email.com', password: 'invalid Password' });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-
-    expect(body.errors).to.have.lengthOf.least(1);
-
-    expect(body.errors[0]).to.have.property('message').eql('Unauthorized. Possible invalid credentials.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('additionalInfo').to.be.eql('User not found.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('code').eq(401);
-
-    expect(body).to.have.nested.property('data').to.be.null;
-  });
-
-  it('Should not login if password does not match.', async () => {
-    const request = QueryLoginMutation({ email: 'useremail1@provider.com', password: 'invalid Password' });
-
-    const { body } = await supertest(baseURL).post('').send(request);
-
-    expect(body.errors).to.have.lengthOf.least(1);
-
-    expect(body.errors[0]).to.have.property('message').eql('Unauthorized. Possible invalid credentials.');
-
-    expect(body.errors[0].extensions.exception)
-      .to.have.property('additionalInfo')
-      .to.be.eql('Password does not match.');
-
-    expect(body.errors[0].extensions.exception).to.have.property('code').eq(401);
-
-    expect(body).to.have.nested.property('data').to.be.null;
-  });
-});
+require('./user-login.mutation.spec');
+require('./user-login-jwt.spec');
